Improve AddGreenEnergyHub performance
The algoritm for finding and registering IHubMessages was O(n^2) or worse.
For every found type, the algoritm had to check the ever growing list of previously found types.

Given that the algorithm checks an unknown number of assemblies with an unknown number of
IHubMessage implementations - this can quickly get out of hand.

Replace the code with an algorithm that runs in O(n).
The code now only handles each type once, as we rely on the dictionary to filter out duplicates
and enforcy priority based on assembly order.