If OperationType is unkown or index of the enum supplied is out of bounds, we no longer throw an exception. (#278)
When we received a OperationType which value was out of bounds, we would throw an exception when determining which rules would apply for the operation. But now we handle the case with the value being unknown OR out of bound with rules that check if the OperationType and the ChargeType is kown.