Adds additional unit tests to charge domain (#408)
This PR adds unit tests to the charge domain by introducing tests that codecov showed we had left untested.

Codecov showed us that we have a number of files that was partial or entirely untested. While a 100% coverage is not the goal, it did reveal a number of files that should have been tested.