Replace sqlite with sqlserver in repository tests (#541)
At the moment we use an SqLite database when running our unit tests. 
We want to replace the implementation to use an SQLServer database instead.

Because we want our test setup to better reflect our production setup. 
And we are experiencing issues with handling time data types.

Technote
This runs tests sequentially in one test class on a single database and 
each test class have their own docker container, so there is no risk of 
having the same database name. But currently we need to clean the 
database between different tests in a testclass if they can affect each other

Story #515