Reject attempt to set TransparentInvoicing true for fee vr904 (#1278)
Due to our Market regulations, the TransparentInvoicing value of a fee must always be false.
A validation rule that checks the provided Transparent Invoicing value is added. In case it is true (for a fee), the transaction will be rejected. This goes for both fee creations and updates.

In addition:
* Add validationrule to documentation
* Add missing check for OperationId in other IValidationRuleForOperation tests
* Add missing check for TriggeredBy in IValidationRuleWithExtendedData tests
* Add missing check for throws exception for invalid ValidationRuleIdentifier in CimValidationErrorCodeFactory