Use FindAsync instead of SingleOrDefaultAsync and add unit test (#1314)
ChargeIngestion and ChargeLinksIngestion both do a Market Participant look-up based on the 'azp' value retrieved from the bearer token. This look-up is using SingleOrDefaultAsync which always does a database round-trip, which can be avoided by using FindAsync, which checks the context for a tracked entity before querying the database (if none found in the context).